+++ changes.md
@@ -4,6 +4,7 @@
 * Fixed the hand card selection screen not wrapping horizontally.
 * Fixed a bug where the character select screen was not properly removed from memory when it was closed, leading to a number of focus issues when attempting to start more than one run in a single session. Again. I think for real this time.
 * Fixed an issue where all card hover tips in buffers were reported as "stolen card".
+* Implemented missing events for Card Played, Potion Used, and End Turn. These default to not reading for the player but reading for other players; this can be configured from event settings.
 
 ## V0.1.7
 * Updated mod manifest to new game format (added `id` field). The mod now works with the latest Slay the Spire 2 update.